HBX-1669: Move class 'org.hibernate.tool.hbmlint.detector.BadCachingDetector' to 'org.hibernate.tool.internal.export.lint.BadCachingDetector'
